Improved op and size handling
PaleTree now saves op name and size to the save file, so we can always
give accurate information about the packets. Previously it always read
the name and size from the current ops, which are incorrect if the ops
change. This is also the case for MabiPale, since we lack a complete op
list there. Only downside: if you have existing log files, they're now
outdated. They will show the correct op, but they will lack the table
size, showing "?".